Fixed Errors => Challenge's Script.
(1) Added checkpoint for each room when you try to admit defeat. If you weren't involved there, then you can't admit defeat.

(2) Fixed the updating problem when you try to update a room (player admits defeat) in a room with a repeated Player1 and Player2 string. Check for only the un-finished room (isDone == false), then apply changes. Fixed the bug that caused the inability to update that room when it's repeated.

(3) No user can create room with other name. This to avoid any problems or cheating trials due to the error that was caused by the bug with the entry (text) (search) not being able to disable that field. o3o